Complete DVLD project

+++ DVLD/DVLD.DAL/DTO/ServiceDTO.cs
@@ -10,15 +10,15 @@ namespace DVLD.DAL.DTO
     {
         public int ID { get; set; }
         public string ServiceName { get; set; }
-        public float ServiceFee { get; set; }
+        public Decimal ServiceFee { get; set; }
 
-        public ServiceDTO(int ID, string ServiceName, float ServiceFee) 
+        public ServiceDTO(int ID, string ServiceName, Decimal ServiceFee) 
         {
             this.ID = ID;
             this.ServiceName = ServiceName;
             this.ServiceFee = ServiceFee;
         }
-        public ServiceDTO(string ServiceName, float ServiceFee)
+        public ServiceDTO(string ServiceName, Decimal ServiceFee)
         {
             this.ServiceName = ServiceName;
             this.ServiceFee = ServiceFee;
